+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Suppress Spotify API HTTP errors from flooding console output
 * Change default volume from 80% to 100%
+* Require Last.fm to be configured and working to start
 
 ### 1.0.10: 2026-01-07